move standardization helpers to different class
